add plugins panel
